requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
spotipy==2.23.0
python-dotenv==1.0.0